better cargo doc by separating by category
