+++ src/rust/rpc-client/src/category/mod.rs
@@ -1,2 +1,7 @@
-mod loan;
-mod vault;
+pub mod loan;
+pub mod misc;
+pub mod oracle;
+pub mod poolpair;
+pub mod token;
+pub mod transaction;
+pub mod vault;
